+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update toggles in admin table to be more helpful
 - Fix filter chips not being un-toggled correctly
 - Changed new date section on Discover to display the current month
+- Fix `Back to Explore` button not respecting filters set by Discover page
 
 ## [v3.3.0](https://github.com/nathanab/beacon/compare/v3.2.0...v3.3.0)
 - Added LogRocket integration